+++ test/deletePath/delete.test.ts
@@ -6,7 +6,7 @@ const table1 = database.s.public.t.table1
 const table2 = database.s.public.t.table2
 
 describe('DELETE Path', () => {
-  const sql = new Builder(database, { allowDeleteWithoutWhereConditions: 'always' })
+  const sql = new Builder(database, { throwErrorIfDeleteHasNoCondition: false })
   afterEach(() => { sql.cleanUp() })
   describe('basic delete all', () => {
     it('Produces [DELETE FROM "table1"]', () => {